+++ DAL/IPrintAccountDB.cs
@@ -2,7 +2,7 @@
 {
     public interface IPrintAccountDB
     {
-        bool AddChfByUsername(string username, double amount);
-        double GetChfByUsername(string username);
+        bool AddChfByUsername(string username, decimal amount);
+        decimal GetChfByUsername(string username);
     }
 }
